+++ x-pack/filebeat/input/httpjson/rate_limiter.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"github.com/elastic/beats/v7/libbeat/logp"
+	"github.com/elastic/beats/v7/x-pack/filebeat/input/httpjson/config"
 )
 
 type rateLimiter struct {
@@ -22,7 +23,7 @@ type rateLimiter struct {
 	remaining string
 }
 
-func newRateLimiterFromConfig(config config, log *logp.Logger) *rateLimiter {
+func newRateLimiterFromConfig(config config.Config, log *logp.Logger) *rateLimiter {
 	if config.RateLimit == nil {
 		return nil
 	}
